+++ db/migrate/20191001222932_update_ae_adverse_event_index_on_number_to_include_project.rb
@@ -1,0 +1,11 @@
+class UpdateAeAdverseEventIndexOnNumberToIncludeProject < ActiveRecord::Migration[6.0]
+  def up
+    remove_index :ae_adverse_events, :number
+    add_index :ae_adverse_events, [:project_id, :number], unique: true
+  end
+
+  def down
+    remove_index :ae_adverse_events, [:project_id, :number]
+    add_index :ae_adverse_events, :number, unique: true
+  end
+end